Refactor.
* Add `gate_model_shared` folder.
* Update `jaqcd.results` and `jaqcd.shared_models` to point to `gate_model_shared` instead.
