+++ src/braket/ir/gate_model_shared/__init__.py
@@ -1,0 +1,10 @@
+from braket.ir.gate_model_shared.results import (  # noqa: F401
+    AdjointGradient,
+    Amplitude,
+    DensityMatrix,
+    Expectation,
+    Probability,
+    Sample,
+    StateVector,
+    Variance,
+)
